Create Ollivander's Inventory - Medium 2
This is the Oracle SQL solution to the following HackerRank challenge: https://www.hackerrank.com/challenges/harry-potter-and-wands/problem

Hermione decides the best way to choose is by determining the minimum number of gold galleons needed to buy each non-evil wand of high power and age. Write a query to print the id, age, coins_needed, and power of the wands that Ron's interested in, sorted in order of descending power. If more than one wand has same power, sort the result in order of descending age.